+++ eslint.config.js
@@ -1,6 +1,7 @@
 import js from '@eslint/js';
 import stylistic from '@stylistic/eslint-plugin';
 import vueTsConfig from '@vue/eslint-config-typescript';
+import prettier from 'eslint-config-prettier/flat';
 import pluginVue from 'eslint-plugin-vue';
 import tseslint from 'typescript-eslint';
 
@@ -19,13 +20,9 @@ export default tseslint.config(
   ...pluginVue.configs['flat/recommended'],
   ...vueTsConfig(),
   stylistic.configs['recommended-flat'],
+  prettier,
   {
     rules: {
-      '@stylistic/semi': ['error', 'always'],
-      '@stylistic/quotes': ['error', 'single', { avoidEscape: true }],
-      '@stylistic/member-delimiter-style': 'off',
-      '@stylistic/indent': ['error', 2],
-      '@stylistic/arrow-parens': ['error', 'always'],
       '@typescript-eslint/consistent-type-imports': ['error', { prefer: 'type-imports' }],
       '@typescript-eslint/no-unused-vars': [
         'error',